gltf: Move JSON writing to json.cpp/write.cpp
json.cpp contains basic JSON helpers, write.cpp contains glTF-specific
bits.
